Create python
IDLE is saying that my return and print statements on lines 52 and 53 have "invalid syntax".  Can't figure out why.  Note that I started writing this on one version of Python and then updated when my computer decided not to run the old version.  Some of the commented sections that I used to check my program originally still have old syntax.